Added ICA-Aroma functionality to preproc.
This might be a creeping feature, but we'll see. ICA denoising is
supposed to be really worth the effort, especially for rs-fmri. And
the aroma tool helps you avoid the pain of training FIX...

Had to pull high-pass filtering out of feat and run it manually to
get it to work, since aroma requires unfiltered data.

Still needs to be tested...