Added block_diag function in npdl_utils
cfd_mat creation started throwing an error in roi_extract, due to a
problem in the Scipy block_diag function. Apparently in a recent update
(v. 0.17.0) it stopped being able to handle input matrices with 0 width.

Added a custom block_diag function to npdl_utils that can handle this
kind of situation.