Changed how roi_extract handles trial durations
roi_extract works with "TR durations". These are trial durations
measured in TRs, relative to the start of the TR containing the trial
onset. Previously, when trials of the same condition had different TR
durations, this raised an error. This error behavior has been removed,
and now instead a warning is given, and the average TR duration is used
for generating the PSC HRF.